Refactor about page layout and add images
This commit refactors the layout of the about page and adds images to enhance the visual experience. The navbar background color has been changed to white. The CallToAction component has been modified to use the CustomButton component, with added support for target="_blank" to open links in a new tab. The DirectionAwareHover component has been updated to remove a console.log statement. The about page now includes two images to provide visual context and enhance the content.

#code-changes
- Refactor navbar background color to white
- Refactor CallToAction and CustomButton components to add target="_blank" support
- Update the layout of the about page to include images